(#20) Split up mouse implementation
